sort: parallel merging
This parallelizes the process of merging files.

It is _not_ useful to be merged, because:
- This only brings improvements when IO is _not_ the bottleneck, but CPU is.
In other words, only SSDs benefit from it.
- This regresses performance on HDD heavily, which requires to only turn
it on when we are sure that this is not the case.
- This is only useful for merging a big number of files, which is
unlikely to be the result of extsort (default file size 100M)
- In some cases CPU usage is increased without big speed improvements
- Memory usage and disk usage is increased

Future work should either find a way to detect whether IO or CPU is the
bottleneck and scale accordingly, or to find a cross-platform way to
reliably detect whether the input and tmp dir is on a flash drive or
a rotational disk.